% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colores.R
\name{obtener_paleta}
\alias{obtener_paleta}
\title{Obtener la paleta de colores}
\usage{
obtener_paleta()
}
\value{
Un vector con los colores definidos.
}
\description{
Esta función devuelve una paleta de colores predefinidos.
}
