% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colores.R
\name{scale_labdem_continuous}
\alias{scale_labdem_continuous}
\title{Paleta de colores personalizada para variables continuas}
\usage{
scale_labdem_continuous()
}
\value{
Un scale_color_gradientn para usar con ggplot2 para variables continuas.
}
\description{
Esta función devuelve un scale_color_gradientn con los colores definidos en la paleta.
}
