% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colores.R
\name{scale_labdem_palette}
\alias{scale_labdem_palette}
\title{Paleta de colores personalizada para ggplot}
\usage{
scale_labdem_palette()
}
\value{
scale_color_manual para usar con ggplot2.
}
\description{
Paleta de colores personalizada para ggplot
}
